Update GCE demo to be more consistent and readable
The IP addresses specified at the start of the demo to be used as reference to the GCE hosts are never actually used in the rest of the demo.  There are 3 different IP values/sets of values referenced in the demo.  These have been modified to be consistent throughout the demo in order to prevent confusion.

There are a few other pieces of instruction that may seem a bit ambiguous.  These have been modified to be more concrete and descriptive so that the user can avoid some trial and error.

CoreOS by default does not require a password.  When networking a docker container and copying config from host-2 to host-1, the CoreOS servers ask for a password in order to complete the file transfer.  The user will not know the password since it was not required to create the machine.  Instead, gcloud compute copy-files can be used.  Added a workaround for this to the instructions.

Gave a name (networking server) to the third GCE VM to make it easier to distinguish which commands to run on which machine.